+++ src/ollama-code-explain.tsx
@@ -1,7 +1,5 @@
-import { ResultView } from "./api/main";
+import { AnswerView } from "./api/ui/AnswerView";
 
 export default function Command(): JSX.Element {
-  const systemPrompt =
-    "Act as a developer. Explain the following code block step by step.\n\nOutput only with the commented code.\n";
-  return ResultView("codeexplain", systemPrompt);
+  return AnswerView("codeexplain");
 }
