+++ src/ollama-fix-spelling-grammar.tsx
@@ -1,7 +1,5 @@
-import { ResultView } from "./api/main";
+import { AnswerView } from "./api/ui/AnswerView";
 
 export default function Command(): JSX.Element {
-  const systemPrompt =
-    "Act as a writer. Fix the following text from spelling and grammar error.\n\nOutput only with the fixed text.\n";
-  return ResultView("fix", systemPrompt);
+  return AnswerView("fix");
 }
