+++ src/ollama-improve-writing.tsx
@@ -1,7 +1,5 @@
-import { ResultView } from "./api/main";
+import { AnswerView } from "./api/ui/AnswerView";
 
 export default function Command(): JSX.Element {
-  const systemPrompt =
-    "Act as a writer. Improve the writing of the following text while keeping the core idea.\n\nOutput only with the modified text.\n";
-  return ResultView("improve", systemPrompt);
+  return AnswerView("improve");
 }
